SAAS-00: Add root screen

+++ lib/data/models/initialise/initialise_app_model.dart
@@ -125,7 +125,7 @@ class BleModule {
 
 class FeatureDetailModel {
   final String? featureKey;
-  final int? featureId;
+  final double? featureId;
   final bool? isStatistics;
   final String? title;
   final String? icon;
